requirements.txt
boto3==1.34.0
pandas==2.2.3
numpy==1.26.4
psycopg2-binary>=2.9.10
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
